yarn -> bun, use TypeScript (miso.ts)
- Port some of miso.js to TypeScript (miso.ts)
- Use bun test runner (no more jsdom / jest)
- Spruce up the .gitignore
- Use bun.lock
- Migrate tests to bun:test
- Address tsc errors w/ miso.ts
- Introduce index.ts (used to generate miso.js for Haskell build)
- Use ESM syntax w/ miso.spec.js
- Drop yarn.lock
- Export top-level miso object namespace
- Intrdouce .tsconfig
