+++ bunfig.toml
@@ -1,0 +1,3 @@
+[test]
+preload = "./ts/happydom.ts"
+coverage = true
