+++ eslint.config.js
@@ -1,25 +1,24 @@
 import { defineConfig } from "eslint/config";
 import globals from "globals";
 import js from "@eslint/js";
-import jestPlugin from "eslint-plugin-jest";
 
 export default defineConfig([
-  { files: ["**/*.{js,mjs,cjs}"] },
-  { files: ["**/*.js"], languageOptions: { sourceType: "script" } },
+  { files: ["**/*.{js,mjs,cjs,ts}"] },
+  { files: ["**/*.js","**/*.ts"], languageOptions: { sourceType: "script" } },
   {
-    files: ["**/*.{js,mjs,cjs}"],
+    files: ["**/*.{js,mjs,cjs,ts}"],
     languageOptions: {
-      globals: { ...globals.browser, ...globals.node, ...globals.jest },
+      globals: { ...globals.browser, ...globals.node },
     },
   },
   {
-    files: ["**/*.{js,mjs,cjs}"],
-    plugins: { js, jestPlugin },
+    files: ["**/*.{js,mjs,cjs,ts}"],
+    plugins: { js },
     extends: ["js/recommended"],
   },
   {
     rules: {
-      "no-global-assign": ["error", { exceptions: ["module"] }],
+      "no-global-assign": ["error", { exceptions: ["miso"] }],
       "arrow-body-style": ["error", "always"],
     },
   },
