+++ package.json
@@ -3,13 +3,15 @@
   "version": "1.9.0",
   "description": "miso: A tasty Haskell front-end web framework",
   "scripts": {
-    "test": "jest --silent",
-    "coverage": "jest --collect-coverage=true --coverage-directory=coverage --silent",
-    "pretty": "prettier --write js/*.js",
-    "lint": "eslint --fix js/*.js"
+    "lint": "eslint --fix ts/*.ts",
+    "test": "bun test --silent js/miso.spec.js",
+    "coverage": "bun test --collect-coverage=true --coverage-directory=coverage --silent",
+    "watch": "tsc ts/miso.ts --watch",
+    "pretty": "prettier --write ts/*.ts",
+    "build": "bun build --outfile=js/miso.js ./ts/index.ts --production --target=browser"
   },
   "type": "module",
-  "main": "js/miso.js",
+  "module": "ts/miso.ts",
   "repository": {
     "type": "git",
     "url": "git+https://github.com/dmjio/miso.git"
@@ -24,31 +26,31 @@
   "bugs": {
     "url": "https://github.com/dmjio/miso/issues"
   },
-  "jest": {
-    "verbose": true
-  },
   "prettier": {
     "singleQuote": true,
     "quoteProps": "preserve"
   },
   "homepage": "https://haskell-miso.org",
   "devDependencies": {
-    "@eslint/js": "^9.23.0",
-    "eslint": "^9.23.0",
-    "eslint-plugin-jest": "^28.11.0",
-    "globals": "^16.0.0",
-    "jest": "29.7.0",
-    "jsdom": "26.0.0",
-    "prettier": "3.5.3"
+    "@happy-dom/global-registrator": "^17.4.4",
+    "prettier": "3.5.3",
+    "@types/bun": "latest"
   },
   "files": [
-    "js/miso.js",
+    "ts/miso.ts",
+    "ts/happydom.ts",
+    "ts/index.ts",
     "js/miso.spec.js",
+    "js/miso.js",
     "package.json",
-    "yarn.lock",
     "README.md"
   ],
   "dependencies": {
-    "jest-environment-jsdom": "^29.7.0"
-  }
+    "eslint": "^9.23.0",
+    "typescript": "^5.8.2"
+  },
+  "peerDependencies": {
+    "typescript": "^5.0.0"
+  },
+  "private": true
 }
