+++ ts/happydom.ts
@@ -1,0 +1,3 @@
+import { GlobalRegistrator } from '@happy-dom/global-registrator';
+
+GlobalRegistrator.register();
