+++ ts/index.ts
@@ -1,0 +1,25 @@
+import
+{ diff
+, hydrate
+, version
+, delegate
+, callBlur
+, callFocus
+, eventJSON
+, undelegate
+, integrityCheck
+, setBodyComponent
+} from './miso';
+
+/* export globally */
+miso = {};
+miso['diff'] = diff;
+miso['hydrate'] = hydrate;
+miso['version'] = version;
+miso['delegate'] = delegate;
+miso['callBlur'] = callBlur;
+miso['callFocus'] = callFocus;
+miso['eventJSON'] = eventJSON;
+miso['undelegate'] = undelegate;
+miso['integrityCheck'] = integrityCheck;
+miso['setBodyComponent'] = setBodyComponent;
